requirements.txt
libconf==1.0.1
paho-mqtt==1.3.1
pyserial==3.4
pyserial-asyncio==0.4
